remove factory reset option

+++ timemachine/main.py
@@ -43,6 +43,9 @@ CLOUD_PATH = "https://storage.googleapis.com/spertilo-data"
 
 def factory_reset():
     print("Reseting to factory settings")
+    if not utils.path_exists("/.factory_lib"):
+        print("Unable to perform factory reset")
+        return
     utils.remove_dir("./lib")
     utils.copy_dir("./factory_lib", "./lib")
     return
@@ -255,7 +258,6 @@ def reconfigure():
             "Test Buttons",
             "Calibrate Knobs",
             "Calibrate Screen",
-            "FactoryReset",
             "Exit",
         ],
     )
